Setup createPages and onCreateNode
In onCreateNode we're adding a slug to the fields of every "MarkdownRemark" node based on its file path.
In createPages we're dynamically creating a page for each markdown file found under "src/data/content".

We also added a React component `BlogTemplate` that will serve as a template for each created page.